move up popup component & update whatsnew
